Removed UIDebugCheats class
