+++ Assets/BossRoom/Prefabs/DebugCheatsManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d026428052f506a4a99ce1b7c94c4d45
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
