+++ Assets/BossRoom/Prefabs/Game/StaticNetworkObjects/GlobalStaticNetworkObjects.prefab
@@ -28,6 +28,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 6803673045227533488}
+  - {fileID: 6432827414648884918}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -104,4 +105,70 @@ PrefabInstance:
 Transform:
   m_CorrespondingSourceObject: {fileID: 5655447013084909147, guid: 8b9c63e7d70c5ff48a03aad51e17103c, type: 3}
   m_PrefabInstance: {fileID: 1159588455107771115}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1001 &8976529767578041044
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 2127969045013110154}
+    m_Modifications:
+    - target: {fileID: 776943915682426285, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_Name
+      value: DebugCheatsManager
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_RootOrder
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 6695920927348029111, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+      propertyPath: GlobalObjectIdHash
+      value: 951099334
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+--- !u!4 &6432827414648884918 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 2726635115303138402, guid: d026428052f506a4a99ce1b7c94c4d45, type: 3}
+  m_PrefabInstance: {fileID: 8976529767578041044}
   m_PrefabAsset: {fileID: 0}
