+++ Assets/BossRoom/Prefabs/UI/BossRoomHudCanvas.prefab
@@ -1399,7 +1399,6 @@ GameObject:
   - component: {fileID: 5277713031684110774}
   - component: {fileID: 5277713031684110769}
   - component: {fileID: 5277713031684110768}
-  - component: {fileID: 9076100988238400401}
   m_Layer: 5
   m_Name: BossRoomHudCanvas
   m_TagString: GameCanvas
@@ -1493,24 +1492,6 @@ MonoBehaviour:
   m_BlockingMask:
     serializedVersion: 2
     m_Bits: 4294967295
---- !u!114 &9076100988238400401
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5277713031684110771}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2dccd56273595344890c4570bcbeb78a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  m_DebugCheatsMediator: {fileID: 0}
-  m_DebugCheatsPanel: {fileID: 7649295862146065024}
-  m_OpenWindowKeyCode: 47
-  m_SpawnEnemyGameEvent: {fileID: 11400000, guid: 157db487e1515e947b4c6e14d99a1fce, type: 2}
-  m_SpawnBossGameEvent: {fileID: 11400000, guid: 4fd8e6b1eeccecd40b18de52f9a6bfb7, type: 2}
-  m_GoToPostGameGameEvent: {fileID: 11400000, guid: e7897c39830fa7644bede6450514147b, type: 2}
 --- !u!1 &5277713032042430138
 GameObject:
   m_ObjectHideFlags: 0
@@ -2581,18 +2562,6 @@ PrefabInstance:
       propertyPath: m_Name
       value: GoToPostGameBtn
       objectReference: {fileID: 0}
-    - target: {fileID: 7721113380319664588, guid: addf3685293c35d48a673397c748f116, type: 3}
-      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_Target
-      value: 
-      objectReference: {fileID: 9076100988238400401}
-    - target: {fileID: 7721113380319664588, guid: addf3685293c35d48a673397c748f116, type: 3}
-      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_MethodName
-      value: GoToPostGame
-      objectReference: {fileID: 0}
-    - target: {fileID: 7721113380319664588, guid: addf3685293c35d48a673397c748f116, type: 3}
-      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_TargetAssemblyTypeName
-      value: Unity.Multiplayer.Samples.BossRoom.Client.UIDebugCheats, Unity.Multiplayer.Samples.BossRoom.Client
-      objectReference: {fileID: 0}
     - target: {fileID: 7721113380319664589, guid: addf3685293c35d48a673397c748f116, type: 3}
       propertyPath: m_Pivot.x
       value: 0.5
@@ -2881,14 +2850,14 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 7562e0bb2a96dfc4996f8ab34f4a0cee, type: 3}
---- !u!1 &5277713031219212455 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 7109257144152377103, guid: 7562e0bb2a96dfc4996f8ab34f4a0cee, type: 3}
-  m_PrefabInstance: {fileID: 3141000975407986600}
-  m_PrefabAsset: {fileID: 0}
 --- !u!224 &5277713031219212452 stripped
 RectTransform:
   m_CorrespondingSourceObject: {fileID: 7109257144152377100, guid: 7562e0bb2a96dfc4996f8ab34f4a0cee, type: 3}
+  m_PrefabInstance: {fileID: 3141000975407986600}
+  m_PrefabAsset: {fileID: 0}
+--- !u!1 &5277713031219212455 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 7109257144152377103, guid: 7562e0bb2a96dfc4996f8ab34f4a0cee, type: 3}
   m_PrefabInstance: {fileID: 3141000975407986600}
   m_PrefabAsset: {fileID: 0}
 --- !u!1001 &3784538605322698414
@@ -3144,18 +3113,6 @@ PrefabInstance:
       propertyPath: m_Name
       value: SpawnBossBtn
       objectReference: {fileID: 0}
-    - target: {fileID: 7721113380319664588, guid: addf3685293c35d48a673397c748f116, type: 3}
-      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_Target
-      value: 
-      objectReference: {fileID: 9076100988238400401}
-    - target: {fileID: 7721113380319664588, guid: addf3685293c35d48a673397c748f116, type: 3}
-      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_MethodName
-      value: SpawnBoss
-      objectReference: {fileID: 0}
-    - target: {fileID: 7721113380319664588, guid: addf3685293c35d48a673397c748f116, type: 3}
-      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_TargetAssemblyTypeName
-      value: Unity.Multiplayer.Samples.BossRoom.Client.UIDebugCheats, Unity.Multiplayer.Samples.BossRoom.Client
-      objectReference: {fileID: 0}
     - target: {fileID: 7721113380319664589, guid: addf3685293c35d48a673397c748f116, type: 3}
       propertyPath: m_Pivot.x
       value: 0.5
@@ -3395,18 +3352,6 @@ PrefabInstance:
     - target: {fileID: 7721113380319664578, guid: addf3685293c35d48a673397c748f116, type: 3}
       propertyPath: m_Name
       value: SpawnImpBtn
-      objectReference: {fileID: 0}
-    - target: {fileID: 7721113380319664588, guid: addf3685293c35d48a673397c748f116, type: 3}
-      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_Target
-      value: 
-      objectReference: {fileID: 9076100988238400401}
-    - target: {fileID: 7721113380319664588, guid: addf3685293c35d48a673397c748f116, type: 3}
-      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_MethodName
-      value: SpawnEnemy
-      objectReference: {fileID: 0}
-    - target: {fileID: 7721113380319664588, guid: addf3685293c35d48a673397c748f116, type: 3}
-      propertyPath: m_OnClick.m_PersistentCalls.m_Calls.Array.data[0].m_TargetAssemblyTypeName
-      value: Unity.Multiplayer.Samples.BossRoom.Client.UIDebugCheats, Unity.Multiplayer.Samples.BossRoom.Client
       objectReference: {fileID: 0}
     - target: {fileID: 7721113380319664589, guid: addf3685293c35d48a673397c748f116, type: 3}
       propertyPath: m_Pivot.x
