+++ Assets/BossRoom/Scenes/BossRoom.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:dbe8dc2c4a754959fecb140560a424f21d3de8290741633620799942260a1d3e
-size 463721
+oid sha256:984855edfd7498d22efbc6bd50666c0a316bb9b19afccb81045f960cbf7242b0
+size 466740
