--- Assets/BossRoom/Scripts/DebugCheats/UIDebugCheats.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2dccd56273595344890c4570bcbeb78a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
